Split out unit tests into two files
